:recycle: [Refactor] noContents 폴더명 변경
